requirements.txt
pycryptodome
cryptography
requests
